requirements.txt
requests
dnspython
argparse
py3_validate_email